add mysqli function to prevent the SQL Injection

+++ admin/script/delete-list-item.php
@@ -1,20 +1,24 @@
 <?php 
 
     include "config.php";
-    $DELETE_ID = $_POST['delete_Id'];
 
-    $query1 = "SELECT * FROM products WHERE id = {$DELETE_ID}";
-    $result1 = mysqli_query($conn, $query1);
-    $row = mysqli_fetch_assoc($result1);
-    unlink("../images/product_img/{$row['img1']}");
-    unlink("../images/product_img/{$row['img2']}");
-    unlink("../images/product_img/{$row['img3']}");
-    unlink("../images/product_img/{$row['img4']}");
+    if($_SERVER['REQUEST_METHOD'] === 'POST'){
+        $DELETE_ID = mysqli_real_escape_string($conn, $_POST['delete_Id']);
+        $query1 = "SELECT * FROM products WHERE id = {$DELETE_ID}";
+        $result1 = mysqli_query($conn, $query1);
+        $row = mysqli_fetch_assoc($result1);
+        unlink("../images/product_img/{$row['img1']}");
+        unlink("../images/product_img/{$row['img2']}");
+        unlink("../images/product_img/{$row['img3']}");
+        unlink("../images/product_img/{$row['img4']}");
 
-    $query = "DELETE FROM products WHERE id = {$DELETE_ID}";
-    $result = mysqli_query($conn, $query);
+        $query = "DELETE FROM products WHERE id = {$DELETE_ID}";
+        $result = mysqli_query($conn, $query);
 
-    if($result && $result1){
-        echo "Product deleted";
+        if($result && $result1){
+            echo "Product deleted";
+        }
+    }else{
+        header("Location: {$host_name}/admin/not-found.php");
     }
 ?>
